chore: abstract getMargin to common tailwind
